beta v1.0.1 (formerly v0.1.0.1)
Similar to before, there's no changelog, but the changes are MASSIVE, and they will all be summarized in the next update.
Changelog in the next update, I swear.

More refactoring! We have a STABLE BUILD!!!
We've completely removed all delimiter-based calculations, and replaced ALL of them with multi-dimensional array lists. This is a continuation on the optimization rounds I was doing in the previous updates. The shift to officially being in beta is because this is the first instance where it is playable with multiple players.

There are some bugs associated with instancing all timers on clients, instead of running them globally;
Users joining mid-round won't have an action bar visible or buff/debuff bar visible. These are visual bugs, so they can still use actions normally. This resolves itself at the end of the round.

*excited scream*