RPGMaker2 beta v1.0.3
The change-log is REALLY massive, so I haven't done it, yet.
This is another MAJOR refactoring update... aka optimization patch. I've gone ahead and rewritten more methods and optimized heavily-used methods such as damage calculators and ability triggers. I've also optimized the common infected, so they should not teleport around as often, and hopefully not at all.
I'll be adding a dps/hps(heals per sec) meter as an add-on to the threat meter in the next update.
After that, I will start re-coding the persistent, equippable ( think Diablo ) loot system, that I had to gut a few patches ago in my roll-back.