v1.0.6a (no version change in code)
Added new key for config.cfg:
"end round if all survivors are incapped?" - 1 to enable, any other value disables.

Now that the class system is gone and players aren't forced into black&white mode if they're a "Defender" build, players who are "Weakened' [Wk] will be set to black&white mode to impress upon them the severity of their situation. I will add a toggle to enable/disable this in a future update.

Fixed a bug where actions equipped to the action bar and not on cooldown weren't providing their passive bonuses to players.

No new optimizations in this patch.