v1.3.5
"ammoreserve" ability trigger with "ammoreseve" result added - this new trigger allows talents to be created that increase a players primary weapon ammo reserves. When picking up ammo from ammo piles, the bonus reserve will be considered and added to the players total inventory.

Added Support Tree layers 1 - 4 to the default talent map found at http://skyrpg.org/current_nodemap.png which is also the default talentmenu.cfg included on this repository.

Corrected a bug where base weapon damage would be increased even if the player had no unlocked or eligible talents in certain categories. Because of this, you will notice the base damage in weapondamages.cfg will now properly show in the character and on the action bar. All player damage across-the-board is lowered to the correct values now, so server operators may need to adjust these values in the weapondamages.cfg to adjust for the increased, but now-accurate TTK.

Added a key to the main config.cfg:
"end round if all survivors are incapped?"				"0"

If set to 0, all survivors have to be dead for the round to end.
However, I made a decision so that if all HUMAN survivor players are dead and the only remaining survivors are bots that the round ends. I will add a switch for this and add it to the config.cfg in another update.

I've also laid the ground-work for a new binary-search-friendly sorting and look-up method for common infected, which will hopefully alleviate any lag issues associated with having too many infected active (and thus making the linear search for the correct common take way too long) and will also have this change in the next version, though you can manually compile with the change if desired.