skyrpg v4.6
skyrpg v4.6 is a stability, optimization, and quality of life update that includes new features for culling super zombie rings/glows, as well as the option to add glows to special infected and other minor features.

Ready Up is still required, but I've put them both in the same folders now for convenience.